slowly moving both album photos and search results into one reusable photo grid component
